backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
django-cors-headers>=4.3
djangorestframework-simplejwt>=5.3
channels>=4.1
channels-redis>=4.2
uvicorn[standard]>=0.30
psycopg2-binary>=2.9
python-dotenv>=1.0
requests>=2.31